xml/parsing/manakai/peref-2.dat

#data
<!DOCTYPE a [
  <!ENTITY % hoge "
    <!ENTITY foo 'hoge foo'>
  ">
]><a>b&foo;c</a>
#errors
1;1;no XML decl;;;s
5;7;entity not declared;;foo;;m
#document
| <!DOCTYPE a>
| <{} a>
|   "b&foo;c"

#data
<!DOCTYPE a [
  <!ENTITY % hoge "
    <!ENTITY foo 'hoge foo'>
  ">
  %hoge;
]><a>b&foo;c</a>
#errors
1;1;no XML decl;;;s
3;5;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a>
| <{} a>
|   "bhoge fooc"

#data
<!DOCTYPE a [
  <!ENTITY % hoge "
    <!ENTITY foo 'hoge <p>foo</p>'>
  ">
  %hoge;
]><a>b&foo;c</a>
#errors
1;1;no XML decl;;;s
3;5;xml:dtd:ext decl;;;w
#document
| <!DOCTYPE a>
| <{} a>
|   "bhoge "
|   <{} p>
|     "foo"
|   "c"

#data
<!DOCTYPE a [
  <!ENTITY % hoge "
    <!ENTITY foo 'hoge <p>foo'>
  ">
  %hoge;
]><a>b&foo;c</a>
#errors
1;1;no XML decl;;;s
3;5;xml:dtd:ext decl;;;w
3;30;in body:#eof;;;m
#document
| <!DOCTYPE a>
| <{} a>
|   "bhoge "
|   <{} p>
|     "foo"
|   "c"

#data
<!DOCTYPE a [
  <!ENTITY % hoge "
    <!ENTITY foo 'hoge <p>foo
  ">
  %hoge;
]><a>b&foo;c</a>
#errors
1;1;no XML decl;;;s
4;3;unclosed entity value;;;m
6;7;entity not declared;;foo;;m
#document
| <!DOCTYPE a>
| <{} a>
|   "b&foo;c"

#data
<!DOCTYPE a [
  <!ENTITY % hoge "
    &#x25;hoge;
    <!ENTITY foo 'hoge <p>foo</p>'>
  ">
  %hoge;
]><a>b&foo;c</a>
#errors
1;1;no XML decl;;;s
3;5;WFC:No Recursion;;%hoge;;m
3;5;stop processing;;;i
4;5;xml:dtd:entity ignored;;;w
7;7;entity not declared;;foo;;m
#document
| <!DOCTYPE a>
| <{} a>
|   "b&foo;c"

#data
<!DOCTYPE a [
  <!ENTITY % hoge "
    &#x25;fuga;
    <!ENTITY foo 'hoge <p>foo</p>'>
  ">
  <!ENTITY % fuga "&#x25;hoge;">
  %hoge;
]><a>b&foo;c</a>
#errors
1;1;no XML decl;;;s
4;5;xml:dtd:entity ignored;;;w
6;20;WFC:No Recursion;;%hoge;;m
6;20;stop processing;;;i
8;7;entity not declared;;foo;;m
#document
| <!DOCTYPE a>
| <{} a>
|   "b&foo;c"

#data
<!DOCTYPE a [
  <!ENTITY % hoge SYSTEM "http://hoge/">
  <!ENTITY % fuga "&#x25;hoge;">
  %hoge;
]><a>b&foo;c</a>
#resource http://hoge/
    %fuga;
    <!ENTITY foo 'hoge <p>foo</p>'>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]2;5;xml:dtd:entity ignored;;;w
3;20;WFC:No Recursion;;%hoge;;m
3;20;stop processing;;;i
5;7;entity not declared;;foo;;m
#document
| <!DOCTYPE a>
| <{} a>
|   "b&foo;c"

#data
<!DOCTYPE a [
  <!ENTITY % hoge SYSTEM "http://hoge/">
  <!ENTITY % fuga "&#x25;hoge;">
  %hoge;
]><a>b&foo;c</a>
#resource http://hoge/
<!ENTITY % abc SYSTEM "http://abc/">
%abc;
#resource http://abc/
    %fuga;
    <!ENTITY foo 'hoge <p>foo</p>'>
#errors
1;1;no XML decl;;;s
[1]1;1;no XML decl;;;s
[1]1;1;xml:dtd:ext decl;;;w
[2]1;1;no XML decl;;;s
[2]2;5;xml:dtd:entity ignored;;;w
3;20;WFC:No Recursion;;%hoge;;m
3;20;stop processing;;;i
5;7;entity not declared;;foo;;m
#document
| <!DOCTYPE a>
| <{} a>
|   "b&foo;c"
